fix: resolve flickering issue in GitHub integration
